[api-rewrite] created api utilities, e.g. publish_check_result
